# backend/docs/swagger.yaml
basePath: /api
definitions:
  handler.AdminResponse:
    description: Admin user data
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        example: admin@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_active:
        example: true
        type: boolean
      last_login:
        example: "2024-01-01T00:00:00Z"
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      username:
        example: admin
        type: string
    type: object
  handler.ChangePasswordSuccessResponse:
    description: Password change success response
    properties:
      message:
        example: password changed successfully
        type: string
    type: object
...
  handler.ErrorResponseBody:
    description: Error response body
    properties:
      error:
        example: something went wrong
        type: string
    type: object
...
  handler.LoginResponse:
    description: Login response with token and admin info
    properties:
      admin:
        $ref: '#/definitions/handler.AdminResponse'
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Mafia Night Support
    url: https://github.com/mafia-night
  description: REST API for the Mafia Night game platform.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Mafia Night API
  version: "1.0"
paths:
  /admin/login:
    post:
      consumes:
      - application/json
      description: Authenticates an admin user and returns a JWT token.
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          properties:
            password:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      summary: Admin login
      tags:
      - admin-auth
...
  /admin/users:
    get:
      description: Retrieves all admin users. Requires admin authentication.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.AdminResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: List admin users
      tags:
      - admin-users
    post:
      consumes:
      - application/json
      description: Creates a new admin user. Requires admin authentication.
      parameters:
      - description: Admin data
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.AdminResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Create an admin user
      tags:
      - admin-users
  /admin/users/{id}:
    delete:
      description: Deletes an admin user. Requires admin authentication.
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Delete an admin user
      tags:
      - admin-users
    get:
      description: Retrieves an admin user by ID. Requires admin authentication.
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AdminResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Get an admin user
      tags:
      - admin-users
    patch:
      consumes:
      - application/json
      description: Updates an admin user. Requires admin authentication.
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: string
      - description: Partial admin data
        in: body
        name: body
        schema:
          properties:
            email:
              type: string
            is_active:
              type: boolean
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AdminResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Update an admin user
      tags:
      - admin-users
  /admin/users/{id}/change-password:
    post:
      consumes:
      - application/json
      description: Changes the password for an admin user. Requires admin authentication.
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: string
      - description: Password change data
        in: body
        name: body
        required: true
        schema:
          properties:
            new_password:
              type: string
            old_password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ChangePasswordSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Change admin password
      tags:
      - admin-users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"